django-ddmi-libs/briderms/requirements.txt
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cx_oracle==7.2.3
Django==3.0.3
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
jinja2==2.11.1
kafka-python==1.4.7
markupsafe==1.1.1
openapi-codec==1.3.2
pycparser==2.19
pyopenssl==19.1.0
pysocks==1.7.1
coreapi==2.3.3
coreschema==0.0.4
requests==2.22.0
setuptools==45.2.0
simplejson==3.17.0
six==1.14.0
tk==0.1.0
uritemplate==3.0.1
urllib3==1.25.8
wheel==0.34.2